% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/better_split.R
\name{better_split}
\alias{better_split}
\title{Better Split}
\usage{
better_split(input, split)
}
\arguments{
\item{input}{the vector of info to parse}

\item{split}{the delimiter}
}
\description{
Parses some demographic data (name, city, state) when delimited by age
}
\examples{
better_split()
}
\keyword{parse}
\keyword{split,}
